fix windows compile and broken aeon
- fix windows linker error during compile
- fix wrong parameter to call aeon (nvidia-backend)
